% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_PFTC_data.R
\name{download_PFTC_data}
\alias{download_PFTC_data}
\title{download PFTC data}
\usage{
download_PFTC_data(country, datatype, path)
}
\arguments{
\item{country}{name of country}

\item{datatype}{name of datatype}

\item{path}{path for downloaded files}
}
\description{
Function to download PFTC data
}
