% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seed.R
\name{get_prefix}
\alias{get_prefix}
\title{get prefix from prompt}
\usage{
get_prefix(prefix = "")
}
\arguments{
\item{prefix}{character Default value for prefix}
}
\value{
Character to use as prefix
}
\description{
get prefix from prompt
}
