% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seed.R
\name{get_seed}
\alias{get_seed}
\title{get seed from prompt}
\usage{
get_seed(seed)
}
\arguments{
\item{seed}{integer Default value for seed}
}
\value{
Integer to use as seed
}
\description{
get seed from prompt
}
