% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PFTC_data_location}
\alias{PFTC_data_location}
\title{PFTC_data_location file}
\format{
A data frame with 32 rows and 6 variables:
\describe{
\item{Country}{country name}
\item{DataType}{data type name}
\item{node}{node name}
\item{file}{file name}
\item{remote_path}{remot path name}
\item{Remark}{remark name}
}
}
\usage{
PFTC_data_location
}
\description{
A dataset containing the data type and location of the data sets
}
\keyword{datasets}
