% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PFTC_envelope_code.R
\name{unique_PFTC_envelope_codes}
\alias{unique_PFTC_envelope_codes}
\title{unique envelope codes across seeds}
\usage{
unique_PFTC_envelope_codes(newseed, oldseeds)
}
\arguments{
\item{newseed}{integer newseed value}

\item{oldseeds}{integer oldseeds value}
}
\value{
tibble with IDs
}
\description{
Function checking that enveloope codes (hashcodes) are unique across seeds. And it selects away non unique codes.
}
\examples{
unique_PFTC_envelope_codes(newseed = 1, oldseeds = 32)
}
