% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PFTC_envelope_code.R
\name{get_PFTC_envelope_codes}
\alias{get_PFTC_envelope_codes}
\title{creates unique envelope codes}
\usage{
get_PFTC_envelope_codes(seed)
}
\arguments{
\item{seed}{integer seed value}
}
\value{
tibble with IDs
}
\description{
Function to create unique hashcodes (Peru: seed = 1; Svalbard: seed = 32; Chile seed = 6)
}
\examples{
get_PFTC_envelope_codes(seed = 1)
}
